test(contentful-export): Add integration test
