+++ package.json
@@ -16,10 +16,11 @@
     "postpublish": "npm run clean",
     "lint": "eslint lib bin/* test",
     "pretest": "npm run lint",
-    "test": "npm run test:cover",
-    "test:cover": "BABEL_ENV=test babel-node ./node_modules/istanbul/lib/cli.js cover test/runner",
-    "test:only": "BABEL_ENV=test babel-node ./test/runner",
-    "test:debug": "BABEL_ENV=test babel-node debug ./test/runner",
+    "test": "npm run test:cover && npm run test:integration",
+    "test:cover": "BABEL_ENV=test babel-node ./node_modules/istanbul/lib/cli.js cover test/runner-unit",
+    "test:only": "BABEL_ENV=test babel-node ./test/runner-unit",
+    "test:integration": "npm run build && BABEL_ENV=test babel-node ./test/runner-integration",
+    "test:debug": "BABEL_ENV=test babel-node debug ./test/runner-unit",
     "test:simulate-ci": "trevor",
     "browser-coverage": "npm run test:cover && opener coverage/lcov-report/index.html",
     "semantic-release": "semantic-release pre && npm publish && semantic-release post",
@@ -67,6 +68,7 @@
     "babel-preset-env": "^1.2.2",
     "babel-template": "^6.23.0",
     "babel-types": "^6.23.0",
+    "contentful-management": "^3.7.1",
     "coveralls": "^2.11.6",
     "cz-conventional-changelog": "^2.0.0",
     "eslint": "^3.18.0",
@@ -76,6 +78,7 @@
     "husky": "^0.13.2",
     "in-publish": "^2.0.0",
     "istanbul": "^1.0.0-alpha.2",
+    "nixt": "^0.5.0",
     "opener": "^1.4.1",
     "require-all": "^2.0.0",
     "rimraf": "^2.5.0",
